Move audit logging and permissions checks into savePolicy()
